Create first draft of custom notations
This allows for numbers of symbols/letters other than 26, allows
for multicharacter symbols/letters, and allows for spaces between
mantissa or exponent or between symbols in the exponent. It doesn't
allow for reversal or no mantissa (those would be easy) or other
bases (that would be harder but still possible). Also letter/cancer
internal inheritance structure is changed.