Update dist and docs (see description for explanation)
This PR is actually already in big trouble with regard to dist
conflicts, so this doesn't dig us into a deeper hole, but it does
stop someone from cloning this branch and being confused about
why it isn't working. Going forward, we don't want to commit changes
to dist in the first place.